requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.10
django-admin-rangefilter==0.11.1
django-import-export==3.2.0
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.0
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
